+++ src/scheduler_min_min.hpp
@@ -13,5 +13,5 @@ class MIN_MIN_Scheduler : public EFT_Scheduler
     MIN_MIN_Scheduler(const common_t *common, simgrid_execs_t &dag);
     ~MIN_MIN_Scheduler();
 
-    std::tuple<simgrid_exec_t *, int, unsigned long> next() override;
+    std::tuple<simgrid_exec_t *, int, double> next() override;
 };
